# openapi/swagger/swagger.yaml
basePath: /
definitions:
...
  main.SwaggerEnvelope:
    properties:
      data: {}
      error:
        type: string
      pagination:
        $ref: '#/definitions/main.SwaggerPagination'
      success:
        type: boolean
    type: object
  main.SwaggerPagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  main.SwaggerRequest:
    additionalProperties: {}
    type: object
...
info:
  contact: {}
  description: Contract for the Nex HTTP API.
  license:
    name: MIT
    url: https://opensource.org/license/mit
  title: Nex API
  version: 1.0.0
paths:
...
  /api/payment/callback/alipay:
    get:
      operationId: payment_callback_alipay_route_get
      parameters:
      - description: Gateway notification parameters
        in: body
        name: body
        schema:
          $ref: '#/definitions/main.SwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      summary: GET /api/payment/callback/alipay
      tags:
      - payment
    post:
      operationId: payment_callback_alipay_route_post
      produces:
      - application/json
      responses:
        "200":
          description: gateway notification result
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      summary: POST /api/payment/callback/alipay
      tags:
      - payment
  /api/payment/callback/mock:
    post:
      operationId: payment_callback_mock_route_post
      produces:
      - application/json
      responses:
        "200":
          description: gateway notification result
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      summary: POST /api/payment/callback/mock
      tags:
      - payment
  /api/payment/callback/wechat:
    get:
      operationId: payment_callback_wechat_route_get
      parameters:
      - description: Gateway notification parameters
        in: body
        name: body
        schema:
          $ref: '#/definitions/main.SwaggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      summary: GET /api/payment/callback/wechat
      tags:
      - payment
    post:
      operationId: payment_callback_wechat_route_post
      produces:
      - application/json
      responses:
        "200":
          description: gateway notification result
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      summary: POST /api/payment/callback/wechat
      tags:
      - payment
...
securityDefinitions:
  ApiTokenAuth:
    in: header
    name: Authorization
    type: apiKey
  CronSecretAuth:
    in: header
    name: Authorization
    type: apiKey
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"